feat(context selection): implement clearing context logic
