+++ src/context-selection/context-selection/context-selection.js
@@ -1,0 +1,25 @@
+import { SelectorBar } from '@dhis2/ui'
+import React from 'react'
+import { CategoryOptionComboSelectorBarItem } from '../category-option-combo-selector-bar-item/index.js'
+import { DataSetSelectorBarItem } from '../data-set-selector-bar-item/index.js'
+import { OrgUnitSetSelectorBarItem } from '../org-unit-selector-bar-item/index.js'
+import { PeriodSelectorBarItem } from '../period-selector-bar-item/index.js'
+import { SectionFilterSelectorBarItem } from '../section-filter-selector-bar-item/index.js'
+import { useClearEntireSelection } from '../use-context-selection.js'
+import useShouldHideClearButton from './use-should-hide-clear-button.js'
+
+export default function ContextSelector() {
+    const hideClearButton = useShouldHideClearButton()
+    const clearEntireSelection = useClearEntireSelection()
+    const onClearSelectionClick = hideClearButton ? undefined : clearEntireSelection
+
+    return (
+        <SelectorBar onClearSelectionClick={onClearSelectionClick}>
+            <DataSetSelectorBarItem />
+            <OrgUnitSetSelectorBarItem />
+            <PeriodSelectorBarItem />
+            <CategoryOptionComboSelectorBarItem />
+            <SectionFilterSelectorBarItem />
+        </SelectorBar>
+    )
+}
