+++ src/context-selection/context-selection/use-should-hide-clear-button.js
@@ -1,0 +1,16 @@
+import { useContextSelection } from '../use-context-selection.js'
+
+export default function useShouldHideClearButton() {
+    const [selection] = useContextSelection()
+
+    if (
+        !selection.orgUnitId &&
+        !selection.periodId &&
+        !selection.categoryOptionComboSelection.length &&
+        !selection.sectionFilter
+    ) {
+        return true
+    }
+
+    return false
+}
